requirements.txt
chromadb
langchain
langchain_community
langchain_ollama
sentence-transformers
ollama
pytest
python-dotenv
spacy
# After installing, run: python -m spacy download en_core_web_sm 